reorganize index
