+++ docs/api/index.rst
@@ -1,20 +1,17 @@
 API Documentation
 =================
 
-GluonNLP Toolkit API includes many useful building blocks for NLP data processing and modeling.
-
-.. toctree::
-   :glob:
-   :maxdepth: 1
-   :caption: Notes
-
-   notes/data_api.rst
-   notes/vocab_emb.rst
-   notes/sequence_sampling.rst
+GluonNLP Toolkit API includes many useful building blocks for NLP data processing and modeling. For
+examples of the API usage, checkout:
 
 .. toctree::
    :maxdepth: 2
-   :caption: Package Reference
+
+   notes
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Packages
 
    vocab
    embedding
