+++ docs/api/notes.rst
@@ -1,0 +1,11 @@
+API Usage Examples
+==================
+
+.. toctree::
+   :glob:
+   :maxdepth: 1
+   :caption: Notes
+
+   notes/data_api.rst
+   notes/vocab_emb.rst
+   notes/sequence_sampling.rst
