--- docs/examples/index.rst
@@ -1,13 +1,0 @@
-Examples
-========
-Here are several end-to-end examples of using GluonNLP Toolkit for modeling tasks.
-
-.. toctree::
-   :maxdepth: 1
-
-   word_embedding/word_embedding.ipynb
-   word_embedding/word_embedding_training.ipynb
-   language_model/language_model.ipynb
-   sentiment_analysis/sentiment_analysis.ipynb
-   nmt/gnmt.ipynb
-   nmt/transformer.ipynb
