+++ docs/examples/machine_translation/nmt/gnmt.py
