+++ docs/examples/machine_translation/nmt/transformer.py
