+++ docs/examples/machine_translation/transformer.png
