+++ docs/index.rst
@@ -13,17 +13,34 @@ research ideas and products based on these models. This toolkit offers four main
 
 
 This toolkit assumes that users have basic knowledge about deep learning and
-NLP. Otherwise, please refer to an introduction course such as 
+NLP. Otherwise, please refer to an introduction course such as
 `Deep Learning---The Straight Dope <http://gluon.mxnet.io/>`_ or
 `Stanford CS224n <http://web.stanford.edu/class/cs224n/>`_.
 
+The toolkit supports the following NLP tasks:
+
+1. :doc:`model_zoo/word_embeddings/index`
+2. :doc:`model_zoo/language_model/index`
+3. :doc:`model_zoo/machine_translation/index`
+4. :doc:`model_zoo/text_classification/index`
+5. :doc:`model_zoo/sentiment_analysis/index`
+6. :doc:`model_zoo/text_generation/index`
+
 .. note::
 
-   This project is at an early stage and is under active development.
-   Please expect that it will be updated frequently. Contributions are welcome.
+   We can be found on `Github <https://github.com/dmlc/gluon-nlp>`_. This project is at an early stage
+   and is under active development, so please expect frequent updates.
+   Check out our `contribution guide <http://gluon-nlp.mxnet.io/how_to/contribute.html>`_
+   to see how you can help.
+
+
+.. hint::
+
+   You can find our the doc for our master development branch `here <http://gluon-nlp.mxnet.io/master/index.html>`_.
+
 
 Installation
-------------------
+------------
 
 GluonNLP relies on the recent version of MXNet. The easiest way to install MXNet
 is through `pip <https://pip.pypa.io/en/stable/installing/>`_. The following
@@ -31,7 +48,7 @@ command installs a nightly built CPU version of MXNet.
 
 .. code-block:: console
 
-   pip install --pre --upgrade mxnet
+   pip install --upgrade mxnet==1.3.0
 
 .. note::
 
@@ -51,7 +68,7 @@ Then install the GluonNLP toolkit by
 
 
 A Quick Example
-----------------
+---------------
 
 Here is a quick example that downloads and creates a word embedding model and then
 computes the cosine similarity between two words.
@@ -69,38 +86,46 @@ computes the cosine similarity between two words.
   <p></p>
 
 
-..
-    .. code-block:: python
-
-..
-       import mxnet as mx
-       import gluonnlp as nlp
-
-..
-       # Create a GloVe word embedding.
-       glove = nlp.embedding.create('glove', source='glove.6B.50d')
-       # Obtain vectors for 'baby' and 'infant' in the GloVe word embedding. 
-       baby, infant = glove['baby'], glove['infant']
-
-..
-       def cos_similarity(vec1, vec2):
-           # Normalize the dot product of two vectors with the L2-norm product.
-           return mx.nd.dot(vec1, vec2) / (vec1.norm() * vec2.norm())
-
-..
-       print(cos_similarity(baby, infant))
-
-
-
 Contents
 --------
 
-.. toctree::
-   :glob:
-   :maxdepth: 1
 
-   self
+.. toctree::
+   :maxdepth: 1
+   :caption: Model Zoo
+
+   model_zoo/word_embeddings/index
+   model_zoo/language_model/index
+   model_zoo/machine_translation/index
+   model_zoo/text_classification/index
+   model_zoo/sentiment_analysis/index
+   model_zoo/text_generation/index
+
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Tutorials
+
+   examples/word_embedding/word_embedding.ipynb
+   examples/word_embedding/word_embedding_training.ipynb
+   examples/language_model/language_model.ipynb
+   examples/sentiment_analysis/sentiment_analysis.ipynb
+   examples/machine_translation/gnmt.ipynb
+   examples/machine_translation/transformer.ipynb
+
+
+.. toctree::
+   :maxdepth: 1
+   :caption: API
+
    api/index
-   examples/index
-   scripts/index
+   api/notes/data_api.rst
+   api/notes/vocab_emb.rst
+   api/notes/sequence_sampling.rst
+
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Community
+
    how_to/contribute
