--- scripts/index.rst
@@ -1,14 +1,0 @@
-Scripts
-=======
-
-.. include:: word_embeddings/word_embeddings.rst
-
-.. include:: sequence_sampling/sequence_sampling.rst
-
-.. include:: language_model/language_model.rst
-
-.. include:: sentiment_analysis/sentiment_analysis.rst
-
-.. include:: nmt/machine_translation.rst
-
-.. include:: classification/document_classification_fasttext.rst
