+++ scripts/sentiment_analysis/index.rst
@@ -1,7 +1,10 @@
-Sentiment Analysis through Fine-tuning, w/ Bucketing
-----------------------------------------------------
+Sentiment Analysis
+------------------
 
-:download:`[Download] </scripts/sentiment_analysis.zip>`
+:download:`[Download] </model_zoo/sentiment_analysis.zip>`
+
+Through Fine-tuning Word Language Model
++++++++++++++++++++++++++++++++++++++++
 
 This script can be used to train a sentiment analysis model from scratch, or fine-tune a pre-trained language model.
 The pre-trained language models are loaded from Gluon NLP Toolkit model zoo. It also showcases how to use different
@@ -11,11 +14,10 @@ Use the following command to run without using pre-trained model (`log <https://
 
 .. code-block:: console
 
-   $ python sentiment_analysis.py --gpu 0 --batch_size 16 --bucket_type fixed --epochs 3 --dropout 0 --no_pretrained --lr 0.005 --valid_ratio 0.1 --save-prefix imdb_lstm_200  # Test Accuracy 85.60
+   $ python finetune_lm.py --gpu 0 --batch_size 16 --bucket_type fixed --epochs 3 --dropout 0 --no_pretrained --lr 0.005 --valid_ratio 0.1 --save-prefix imdb_lstm_200  # Test Accuracy 85.60
 
 Use the following command to run with pre-trained model (`log <https://github.com/dmlc/web-data/blob/master/gluonnlp/logs/sentiment/sentiment_pretrained_20180817.log>`__)
 
 .. code-block:: console
 
-   $ python sentiment_analysis.py --gpu 0 --batch_size 16 --bucket_type fixed --epochs 3 --dropout 0 --lr 0.005 --valid_ratio 0.1 --save-prefix imdb_lstm_200  # Test Accuracy 86.46
-
+   $ python finetune_lm.py --gpu 0 --batch_size 16 --bucket_type fixed --epochs 3 --dropout 0 --lr 0.005 --valid_ratio 0.1 --save-prefix imdb_lstm_200  # Test Accuracy 86.46
