+++ scripts/tests/test_encoder_decoder.py
@@ -1,8 +1,8 @@
 import numpy as np
 import mxnet as mx
 from mxnet.test_utils import assert_almost_equal
-from ..nmt.gnmt import *
-from ..nmt.transformer import *
+from ..machine_translation.gnmt import *
+from ..machine_translation.transformer import *
 
 
 def test_gnmt_encoder():
