+++ scripts/tests/test_question_answering.py
@@ -22,7 +22,7 @@ import pytest
 from mxnet.gluon.data import DataLoader, SimpleDataset
 
 from gluonnlp.data import SQuAD
-from scripts.question_answering.data_processing import SQuADTransform, VocabProvider
+from ..question_answering.data_processing import SQuADTransform, VocabProvider
 
 question_max_length = 30
 context_max_length = 256
